+++ packages/apollos-data-connector-rock/src/features/data-source.js
@@ -7,15 +7,21 @@ export default class Feature extends RockApolloDataSource {
   resource = '';
 
   // Names of Action Algoritms mapping to the functions that create the actions.
-  ACTION_ALGORITHIMS = {
+  ACTION_ALGORITHIMS = Object.entries({
     // We need to make sure `this` refers to the class, not the `ACTION_ALGORITHIMS` object.
-    PERSONA_FEED: this.personaFeedAlgorithm.bind(this),
-    CONTENT_CHANNEL: this.contentChannelAlgorithm.bind(this),
-    SERMON_CHILDREN: this.sermonChildrenAlgorithm.bind(this),
-    UPCOMING_EVENTS: this.upcomingEventsAlgorithm.bind(this),
-    CAMPAIGN_ITEMS: this.campaignItemsAlgorithm.bind(this),
-    USER_FEED: this.userFeedAlgorithm.bind(this),
-  };
+    PERSONA_FEED: this.personaFeedAlgorithm,
+    CONTENT_CHANNEL: this.contentChannelAlgorithm,
+    SERMON_CHILDREN: this.sermonChildrenAlgorithm,
+    UPCOMING_EVENTS: this.upcomingEventsAlgorithm,
+    CAMPAIGN_ITEMS: this.campaignItemsAlgorithm,
+    SERIES_IN_PROGRESS: this.seriesInProgressAlgorithm,
+    USER_FEED: this.userFeedAlgorithm,
+  }).reduce((accum, [key, value]) => {
+    // convenciance code to make sure all methods are bound to the Features dataSource
+    // eslint-disable-next-line
+    accum[key] = value.bind(this);
+    return accum;
+  }, {});
 
   getFromId(args, id) {
     const type = id.split(':')[0];
@@ -63,6 +69,55 @@ export default class Feature extends RockApolloDataSource {
       subtitle,
       // Typanme is required so GQL knows specifically what Feature is being created
       __typename: 'ActionListFeature',
+    };
+  }
+
+  async createHeroListFeature({
+    algorithms = [],
+    heroAlgorithms = [],
+    title,
+    subtitle,
+  }) {
+    // Generate a list of actions.
+    let actions;
+    let heroCard;
+    // If we have a strategy for selecting the hero card.
+    if (heroAlgorithms && heroAlgorithms.length) {
+      // The actions come from the action algorithms
+      actions = () => this.runAlgorithms({ algorithms });
+      // and the hero comes from the hero algorithms
+      heroCard = async () => {
+        const cards = await this.runAlgorithms({ algorithms: heroAlgorithms });
+        return cards.length ? cards[0] : null;
+      };
+      // Otherwise, if we don't have a strategy
+    } else {
+      // Get all the cards (sorry, no lazy loading here)
+      const allActions = await this.runAlgorithms({ algorithms });
+      // The actions are all actions after the first
+      actions = allActions.slice(1);
+      // And the hero is the first action.
+      heroCard = allActions.length ? allActions[0] : null;
+    }
+
+    return {
+      // The Feature ID is based on all of the action ids, added together.
+      // This is naive, and could be improved.
+      id: this.createFeatureId({
+        type: 'HeroListFeature',
+        args: {
+          algorithms,
+          heroAlgorithms,
+          title,
+          subtitle,
+        },
+      }),
+      actions,
+      heroCard,
+      title,
+      subtitle,
+      // Typanme is required so GQL knows specifically what Feature is being created
+      __typename: 'HeroListFeature',
     };
   }
 
@@ -291,6 +346,25 @@ Make sure you structure your algorithm entry as \`{ type: 'CONTENT_CHANNEL', aru
     }));
   }
 
+  async seriesInProgressAlgorithm({ limit = 3 } = {}) {
+    const { ContentItem } = this.context.dataSources;
+
+    const items = await (await ContentItem.getSeriesWithUserProgress())
+      .expand('ContentChannel')
+      .top(limit)
+      .get();
+
+    return items.map((item, i) => ({
+      id: createGlobalId(`${item.id}${i}`, 'ActionListAction'),
+      title: item.title,
+      subtitle: get(item, 'contentChannel.name'),
+      relatedNode: { ...item, __type: ContentItem.resolveType(item) },
+      image: ContentItem.getCoverImage(item),
+      action: 'READ_CONTENT',
+      summary: ContentItem.createSummary(item),
+    }));
+  }
+
   async getScriptureShareMessage(ref) {
     const { Scripture } = this.context.dataSources;
     const scriptures = await Scripture.getScriptures(ref);
@@ -310,6 +384,8 @@ Make sure you structure your algorithm entry as \`{ type: 'CONTENT_CHANNEL', aru
             return this.createVerticalCardListFeature(featureConfig);
           case 'HorizontalCardList':
             return this.createHorizontalCardListFeature(featureConfig);
+          case 'HeroListFeature':
+            return this.createHeroListFeature(featureConfig);
           case 'ActionList':
           default:
             // Action list was the default in 1.3.0 and prior.