+++ packages/apollos-data-schema/src/index.js
@@ -696,6 +696,16 @@ export const featuresSchema = gql`
     actions: [ActionListAction]
   }
 
+  type HeroListFeature implements Feature & Node {
+    id: ID!
+    order: Int
+
+    title: String
+    subtitle: String
+    actions: [ActionListAction]
+    heroCard: CardListItem
+  }
+
   type CardListItem {
     id: ID!
 
@@ -741,6 +751,15 @@ export const featuresSchema = gql`
     order: Int
 
     scriptures: [Scripture]
+  }
+
+  type WebviewFeature implements Feature & Node {
+    id: ID!
+    order: Int
+
+    url: String
+    title: String
+    linkText: String
   }
 
   extend type WeekendContentItem {